Use shared components for export dialog
    - use dialog and checkbox components
    - little refactoring

Ticket: #247
